Remove control server adaptive script due to deprecation and refactoring of control logic.
